feat: Separate the entry request document from the entry request model.
